Put simdjson routines instead of old-json
